Update ticksPerPound to double in IDA100 class

+++ load_cell/ida100.h
@@ -12,7 +12,7 @@ class IDA100 {
    private:
     const int CALIBRATE_SAMPLES = 50;
 
-    const int ticksPerPound;
+    const double ticksPerPound;
 
     FT_HANDLE ftHandle;
     int calibZero;
